+++ components/02-elements/form/file-upload/file-upload.js
@@ -1,11 +1,30 @@
-'use strict'
+'use strict';
 
-const inputFile = document.querySelector('.file-upload__field');
+(function FileUpload() {
+  const components = [...document.querySelectorAll('.file-upload')];
 
-inputFile.addEventListener('change', (file) => {
-  const filePath          = file.target.value,
-        splitedPath       = filePath.split('\\'),
-        splitedPathLength = splitedPath.length - 1;
+  components.forEach(component => {
+    const input = component.querySelector('.file-upload__input'),
+          name   = component.querySelector('.file-upload__name'),
+          button = component.querySelector('span[role="button"]');
 
-  inputFile.parentNode.querySelector('.file-upload__name').innerHTML = splitedPath[splitedPathLength];
-});
+    input.addEventListener('change', file => {
+      const filePath = file.target.value,
+            fileName = filePath.split('\\').pop();
+
+      if (fileName) {
+        name.textContent = fileName;
+      }
+      else {
+        name.textContent = name.dataset.defaultText;
+      }
+      name.focus();
+    });
+
+    button.addEventListener('keypress', event => {
+      if (event.which === 32 || event.which === 13) {
+        input.click();
+      }
+    });
+  });
+})();